Refactor board offset coordinates in board class
Changes the way the board is drawn with offset. Board.draw(),
Board.draw_tiles() and Board.draw_pieces() used to take 2 with the x
and y offset of the board in pixels. These arguments are now removed
and replaced by 2 private class variables x_offset and y_offset for
cleaner code.